붕대 애니메이션 수정 및 아이템 0개일 때 사용 못하게 막음

+++ Assets/02_Scripts/Item/ItemManager.cs
@@ -7,7 +7,7 @@ public abstract class ItemManager : MonoBehaviour
     [SerializeField, Header("아이템")]
     protected GameObject item;
     [SerializeField, Header("사용 시간")]
-    protected int useTime;
+    protected float useTime;
     [SerializeField, Header("기본 가위 무기")]
     protected GameObject baseWeapon;
